@@ -1,12 +1,12 @@
 package com.web.stard.domain.board.global.domain;
 
-import com.sun.istack.NotNull;
 import com.web.stard.domain.board.global.domain.enums.PostType;
 import com.web.stard.domain.member.domain.Member;
 import com.web.stard.global.domain.BaseEntity;
+import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 @Table
