@@ -1,14 +1,15 @@
 package com.web.stard.domain.board.global.domain;
 
-import com.sun.istack.NotNull;
+
 import com.web.stard.domain.board.global.domain.enums.ActType;
 import com.web.stard.domain.board.global.domain.enums.PostType;
 import com.web.stard.domain.board.study.domain.Study;
 import com.web.stard.domain.board.study.domain.StudyPost;
 import com.web.stard.domain.member.domain.Member;
+import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 @Table
@@ -40,7 +41,8 @@ public class StarScrap {
     @Enumerated(EnumType.STRING)
     private PostType tableType; // 게시글 타입 (COMM / STUDY, STUDYPOST)
 
-    @NotNull @ManyToOne
+    @NotNull
+    @ManyToOne
     @JoinColumn(name = "member_id")
     private Member member;
 }
