@@ -3,7 +3,7 @@ package com.web.stard.domain.board.study.domain;
 import com.web.stard.domain.member.domain.Member;
 import lombok.*;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 @Table
