@@ -1,10 +1,10 @@
 package com.web.stard.domain.board.study.domain;
 
 import com.web.stard.domain.member.domain.Member;
+import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
-import javax.persistence.*;
-import javax.validation.constraints.NotNull;
+import jakarta.persistence.*;
 
 @Entity
 @Builder
