@@ -2,13 +2,13 @@ package com.web.stard.domain.chat_stomp.domain;
 
 import com.web.stard.global.domain.BaseEntity;
 import com.web.stard.domain.member.domain.Member;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import javax.persistence.*;
-import javax.validation.constraints.NotNull;
+import jakarta.persistence.*;
 
 @Getter
 @Setter
