@@ -3,6 +3,7 @@ package com.web.stard.domain.admin.api;
 import com.web.stard.domain.board.global.domain.Post;
 import com.web.stard.domain.member.application.MemberService;
 import com.web.stard.domain.admin.application.NoticeService;
+import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
@@ -11,7 +12,6 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
-import javax.transaction.Transactional;
 import java.util.Optional;
 
 @Transactional
