@@ -1,34 +1,23 @@
 package com.web.stard.global.handler;
 
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.web.authentication.SavedRequestAwareAuthenticationSuccessHandler;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
 
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
-public class LoginSuccessHandler extends SavedRequestAwareAuthenticationSuccessHandler {
-    public LoginSuccessHandler(String defaultTargetUrl) {
-        setDefaultTargetUrl(defaultTargetUrl);
-    }
+public class LoginSuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
 
     @Override
     public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response,
                                         Authentication authentication) throws ServletException, IOException {
 
-        HttpSession session = request.getSession();
-        if (session != null) {
-            String redirectUrl = (String) session.getAttribute("prevPage");
-            if (redirectUrl != null) {
-                session.removeAttribute("prevPage");
-                getRedirectStrategy().sendRedirect(request, response, redirectUrl);
-            } else {
-                super.onAuthenticationSuccess(request, response, authentication);
-            }
-        } else {
-            super.onAuthenticationSuccess(request, response, authentication);
-        }
+        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+        logger.info("로그인 성공. JWT 발급. username: {}" + userDetails.getUsername());
+
+        response.getWriter().write("success");
     }
 }
